More linux warning removals

+++ components/common/cpp/src/loguru.cpp
@@ -510,6 +510,7 @@ namespace loguru
 
 	Text errno_as_text()
 	{
+		// cppcheck-suppress unusedVariable
 		char buff[256];
 	#if defined(__GLIBC__) && defined(_GNU_SOURCE)
 		// GNU Version
@@ -1069,7 +1070,7 @@ namespace loguru
 						 i - skip, int(2 + sizeof(void*) * 2), callstack[i],
 						 status == 0 ? demangled :
 						 info.dli_sname == 0 ? symbols[i] : info.dli_sname,
-						 static_cast<char*>(callstack[i]) - static_cast<char*>(info.dli_saddr));
+						 (ssize_t)(static_cast<char*>(callstack[i]) - static_cast<char*>(info.dli_saddr)));
 				free(demangled);
 			} else {
 				snprintf(buf, sizeof(buf), "%-3d %*p %s\n",
